Remove the ascii art.
The effect is lost when running multiple containers with compose,
as the messages from different containers can interleave.

Instead print a text line with the name of the company.

Also tweak some of the other messages to print on a single
line so they are easier to read if interleaving happens.